♻️ Refactor name file
